requirements.txt
bson
pymongo
genson
dnspython
python_dotenv
python_dateutil